refactor: dashboard-page-buttons-utils.js -> ...-creating-utils.js
