+++ public/js/dashboard/dashboard-utils/dashboard-pages-utils.js
@@ -1,7 +1,7 @@
 import { getTableFromDashboardPage, getContext } from '../dashboard-driver.js';
 import setActivePage from './set-active-page.js';
 import createDashboardItems from '../create-dashboard-items.js';
-import { getPageButtonSpec } from './dashboard-page-buttons-utils.js';
+import { getPageButtonSpec } from './dashboard-page-buttons-creating-utils.js';
 import buildDOM from '../../table/build-dom.js';
 
 /**
